+++ src/syntax/parsing/mod.rs
@@ -1,0 +1,9 @@
+//! Parsing of source code into syntax trees.
+
+mod escaping;
+mod parser;
+
+pub use parser::parse;
+
+#[cfg(test)]
+mod tests;
